Rewrote translation and collapse.
+ major bug in expression evaluation order of operations
+ compiler not attaching attributes to types or functions YET
+ compiler not actually emitting binary form of byte-code YET
+ old code needs to be removed from rewrite of translation and collapse

So, the translation and collapse was re-written (near the top) and
name translation_v2. The old code got messy and wild to read. The
new code uses some object oriented behavior to clean that up and
also proper creates a list that allows evaluating items from the
very edges/tips of the tree down to the root which simplifies doing
the translating and most especially... the collapsing of the tree.