Removed old code in compiler.py
The old code for the first write of the compiler was removed. Some
of the old code was left. This old code that was left included just
a few utility functions and `translate_seq`; however, `translate_seq`
is about to be re-written.

The `translate_seq` is the workhorse by emitting byte-code representing
the evaluation of expressions. Since a program is essentially mostly
expressions along with maybe a store and usually loads.

The major control flow from loops, ifs, and such are provided by the
actual token classes and these emit the needed glue code. Even the
conditions for loops and if statements are effectively expressions so
it makes sense to have something just for translating those into
bytecode.

The current version, in this commit, of `translate_seq` works well
except it does not deal with order of operations and it most especially
produces wrong byte-code when dealing with various boolean type
operations and such.